+++ lib/membrane/filter.ex
@@ -66,7 +66,8 @@ defmodule Membrane.Filter do
 
       @behaviour unquote(__MODULE__)
 
-      @impl true
+      @doc false
+      @spec membrane_element_type() :: Membrane.Element.type_t()
       def membrane_element_type, do: :filter
 
       @impl true