+++ lib/membrane/source.ex
@@ -30,7 +30,8 @@ defmodule Membrane.Source do
       use Membrane.Element.Base, unquote(options)
       use Membrane.Element.WithOutputPads
 
-      @impl true
+      @doc false
+      @spec membrane_element_type() :: Membrane.Element.type_t()
       def membrane_element_type, do: :source
     end
   end