+++ test/membrane/integration/timer_test.exs
@@ -68,4 +68,25 @@ defmodule Membrane.Integration.TimerTest do
     assert_receive :pipeline_tick
     Testing.Pipeline.terminate(pipeline, blocking?: true)
   end
+
+  defmodule StopNoInterval do
+    use Membrane.Source
+    @impl true
+    def handle_setup(_ctx, state) do
+      Process.send_after(self(), :stop_timer, 0)
+      {{:ok, start_timer: {:timer, :no_interval}}, state}
+    end
+
+    @impl true
+    def handle_info(:stop_timer, _ctx, state) do
+      {{:ok, stop_timer: :timer, notify_parent: :ok}, state}
+    end
+  end
+
+  test "Stopping timer with `:no_interval`" do
+    pipeline = Testing.Pipeline.start_link_supervised!(children: [element: StopNoInterval])
+    assert_pipeline_play(pipeline)
+    assert_pipeline_notified(pipeline, :element, :ok)
+    Testing.Pipeline.terminate(pipeline, blocking?: true)
+  end
 end